fix(workflow): serialize canonical backfill writes
